+++ src/main/kotlin/cloud/glitchdev/rfu/feature/mob/GdragAlert.kt
@@ -37,7 +37,7 @@ object GdragAlert : Feature {
                         alertedEntities.add(entity)
                         Title.showTitle("§c§lNO G-DRAGON!", "§eEquip your Golden Dragon!", fadeIn = 5, duration = 40, fadeOut = 5)
                         if (SeaCreatureConfig.goldenDragonSound) {
-                            Sounds.playSound("rfu:gdrag_alert", 1f, 1f)
+                            Sounds.playSound("rfu:gdrag_alert", 1f, SeaCreatureConfig.goldenDragonVolume)
                         }
                     }
                 }